spawning snake head

+++ Linked-List-Snake/source/Level/LevelModel.cpp
@@ -6,7 +6,7 @@ namespace Level
 
 	LevelModel::~LevelModel() = default;
 
-	void LevelModel::initialize(int width, int height)
+	void LevelModel::initialize(float width, float height)
 	{
 		cell_width = width / number_of_columns;
 		cell_height = height / number_of_rows;
